Boards-STM32: use size helpers to describe size of storage partition
The goal of this commit is to update existing STM32 boards descriptions
to use these size "DT_SIZE" macros to enhance readability. To realize this
i used a python script, which will detect the STM32 Boards /zephyr/board,
and then will update in the dts files the partitiondescripitions using
"DT_SIZE_K" and "DT_SIZE_M".
Check manually and modify in .overlay files in samples and tests.


Signed-off-by: TLIG Dhaou <dhaou.tlig-ext@st.com>